+++ src/lib/System.cc
@@ -36,8 +36,4 @@ namespace es {
     // nothing by default
   }
 
-  void System::updateEntity(float delta, Entity entity) {
-    // nothing by default
-  }
-
 }