+++ ChangeLog.md
@@ -1,5 +1,13 @@
 # ChangeLog for `libes`
 
+## `libes` 0.5
+
+### 0.5.0 (29 Jan 2014)
+
+* add an overload for `registerHandler`
+* remove updateEntity from System (put it directly in LocalSystem and GlobalSystem)
+* add CustomSystem and SingleSystem
+
 ## `libes` 0.4
 
 ### 0.4.0 (19 Dec 2013)